Fix "no path" bug for uncoupled wagons.
When wagons are uncoupled from the front of a train, they are left in automatic mode and continuously give a "no path" floating text message.  In addition to forcing segments to automatic when locomotives are present, we also need to force segments to manual when there are none.  This update adds that function and eliminates the "no path" messages for uncoupled wagons.